# prism/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API gateway with service routing, audit logging, API key management, and service discovery.",
        "title": "Prism API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/routes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all registered routes, optionally filtered by type (public or protected)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "List all routes",
                "parameters": [
                    {
                        "enum": [
                            "public",
                            "protected"
                        ],
                        "type": "string",
                        "description": "Filter by route type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RouteMetadata"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/api/routes/{service}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all registered routes for a specific service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "List routes for a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "service",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RouteMetadata"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new custom route for a specific service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Add a route to a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "service",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Route configuration",
                        "name": "route",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RouteConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RouteConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/api/services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all registered backend services",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "List all services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ServiceConfig"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new backend service with the gateway",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Register a new service",
                "parameters": [
                    {
                        "description": "Service configuration",
                        "name": "service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ServiceConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ServiceConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/api/services/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve configuration for a specific registered service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Get a service by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceConfig"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update configuration for an existing service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Update a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated service configuration",
                        "name": "service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ServiceConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deregister a service from the gateway",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Delete a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid JSON request"
                },
                "slug": {
                    "type": "string",
                    "example": "invalid-json"
                }
            }
        },
...
        "Duration": {
            "type": "integer",
            "format": "int64",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        },
...
        "RouteConfig": {
            "type": "object",
            "properties": {
                "method": {
                    "description": "HTTP method (GET, POST, etc.), or * for any",
                    "type": "string"
                },
                "path": {
                    "description": "Path pattern (e.g., \"/api/users/:id\")",
                    "type": "string"
                },
                "priority": {
                    "description": "Higher priority routes are matched first",
                    "type": "integer"
                },
                "public": {
                    "description": "Whether this route is public (no auth required)",
                    "type": "boolean"
                },
                "requiredScopes": {
                    "description": "Required permission scopes (if any)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "RouteMetadata": {
            "type": "object",
            "properties": {
                "method": {
                    "description": "HTTP method (GET, POST, etc.)",
                    "type": "string"
                },
                "path": {
                    "description": "URL path pattern",
                    "type": "string"
                },
                "priority": {
                    "description": "Higher priority routes are matched first",
                    "type": "integer"
                },
                "public": {
                    "description": "true = no auth required, false = auth required",
                    "type": "boolean"
                },
                "required_scopes": {
                    "description": "Required permission scopes (if any)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "service_id": {
                    "description": "ID of the service that handles this route",
                    "type": "string"
                },
                "service_name": {
                    "description": "Name of the service",
                    "type": "string"
                },
                "service_url": {
                    "description": "Base URL of the service",
                    "type": "string"
                },
                "tags": {
                    "description": "Additional metadata tags",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "ServiceConfig": {
            "type": "object",
            "properties": {
                "allowedMethods": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowedRoles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "healthCheckPath": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "requiresAuth": {
                    "type": "boolean"
                },
                "retryCount": {
                    "type": "integer"
                },
                "routes": {
                    "description": "Optional custom routes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RouteConfig"
                    }
                },
                "timeout": {
                    "$ref": "#/definitions/Duration"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}